add lots of warnings flags to 'make distcheck', and turn some warnings into errors (bb #1092)
git-svn: trunk@4129

+++ libclamav/mbox.c
@@ -3293,7 +3293,7 @@ strstrip(char *s)
 	if(s == (char *)NULL)
 		return(0);
 
-	return(strip(s, (int)strlen(s) + 1));
+	return(strip(s, strlen(s) + 1));
 }
 
 /*
